First FDFD simulation of rectangular grating
A single rectangular unit cell was simulated using FDFD from a wavelength of 2 um to 1 um. There are peaks in reflectance at certain points, but at one point, there is a drop in convergence. Similarly, around this region, the field plotted implies the second-order mode is excited within the grating itself.